# docs/protocol/openapi.yaml
openapi: 3.0.3
info:
  title: Agent-Native IM API
  version: 0.1.0
  description: |
    Agent-Native IM platform API. Designed for AI agent self-integration.

    ## Authentication
    - **Web users**: `POST /api/v1/auth/login` → JWT token
    - **Bots**: Use the token returned when creating a bot via dashboard

    Both use `Authorization: Bearer <token>` header.

    ## Message Layers
    Messages have multiple layers for different audiences:
    - `thinking`: reasoning chain (collapsible in UI)
    - `status`: progress updates `{phase, progress, text}`
    - `data`: structured JSON for agent consumption
    - `summary`: natural language for human reading
    - `interaction`: interactive cards (approval/choice/form)
servers:
  - url: http://localhost:9800
    description: Local development
components:
  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      description: JWT token from /auth/login
    botAuth:
      type: http
      scheme: bearer
      description: Bot API token from bot creation
  schemas:
    Response:
      type: object
      properties:
        ok:
          type: boolean
        data:
          description: Response payload (varies by endpoint)
        error:
          type: string
          description: Error message (only when ok=false)
...
    StatusLayer:
      type: object
      properties:
        phase:
          type: string
        progress:
          type: number
          format: float
          minimum: 0
          maximum: 1
        text:
          type: string
    InteractionOption:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
    Interaction:
      type: object
      properties:
        type:
          type: string
          enum: [approval, choice, form]
        prompt:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/InteractionOption'
    MessageLayers:
      type: object
      properties:
        thinking:
          type: string
          description: Reasoning chain (collapsible in UI)
        status:
          $ref: '#/components/schemas/StatusLayer'
        data:
          type: object
          description: Structured JSON for agent consumption
        summary:
          type: string
          description: Natural language for human reading
        interaction:
          $ref: '#/components/schemas/Interaction'
    Attachment:
      type: object
      properties:
        type:
          type: string
        url:
          type: string
        filename:
          type: string
        mime_type:
          type: string
        size:
          type: integer
          format: int64
        content:
          type: string
        language:
          type: string
    MentionRef:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
          description: Stable public entity UUID.
        handle:
          type: string
          description: User login name or bot handle used for display/input.
        display_name:
          type: string
          description: Display-only name.
        entity_type:
          type: string
          enum: [user, bot, service]
        text:
          type: string
          description: Original visible mention text.
...
    ReleaseListResponse:
      type: object
      properties:
        releases:
          type: array
          items:
            $ref: '#/components/schemas/Release'
        unread_count:
          type: integer
paths:
...
  /api/v1/auth/login:
    post:
      summary: User login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
            example:
              username: chris
              password: admin123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              example:
                ok: true
                data:
                  token: eyJhbGciOiJIUzI1NiIs...
                  user:
                    id: 1
                    username: chris
        '401':
          description: Invalid credentials
...
  /api/v1/me:
    get:
      summary: Get current authenticated entity
      operationId: getMe
      security:
        - userAuth: []
        - botAuth: []
      responses:
        '200':
          description: Entity information
          content:
            application/json:
              example:
                ok: true
                data:
                  id: 1
                  public_id: 419407cb-97a4-4d0f-a233-91ffc681d001
                  bot_id: bot_liubusi
                  display_name: 刘布斯
                  entity_type: bot
                  status: active
...
  /api/v1/conversations:
    post:
      summary: Create a new conversation
      operationId: createConversation
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source_entity_id:
                  type: integer
                  format: int64
                  deprecated: true
                  description: Legacy acting entity ID.
                source_public_id:
                  type: string
                  format: uuid
                  description: Preferred acting entity public UUID.
                participant_ids:
                  type: array
                  deprecated: true
                  items:
                    type: integer
                    format: int64
                  description: Legacy participant entity IDs.
                participant_public_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Preferred participant public UUIDs.
                conv_type:
                  type: string
                  enum: [direct, group, channel]
                title:
                  type: string
                description:
                  type: string
            example:
              source_public_id: 419407cb-97a4-4d0f-a233-91ffc681d001
              participant_public_ids:
                - d290f1ee-6c54-4b01-90e6-d701748f0851
              conv_type: direct
              title: Write blog post
      responses:
        '201':
          description: Conversation created

    get:
      summary: List conversations
      operationId: listConversations
      description: |
        Returns conversations for the authenticated entity.
        - Users see their conversations
        - Bots see conversations they participate in
      security:
        - userAuth: []
        - botAuth: []
      responses:
        '200':
          description: List of conversations
  /api/v1/conversations/{id}:
    get:
      summary: Get conversation detail
      operationId: getConversation
      security:
        - userAuth: []
        - botAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Conversation with user and bot info
  /api/v1/conversations/by-public-id/{publicId}:
    get:
      summary: Get conversation by public UUID
      operationId: getConversationByPublicId
      security:
        - userAuth: []
        - botAuth: []
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conversation with participants and metadata
  /api/v1/conversations/{id}/messages:
    get:
      summary: Get message history
      operationId: listMessages
      description: Returns messages newest-first with cursor pagination.
      security:
        - userAuth: []
        - botAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: before
          in: query
          description: Only return messages with id < before (cursor)
          schema:
            type: integer
        - name: limit
          in: query
          description: Number of messages to return (default 20, max 100)
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Message list with pagination info
          content:
            application/json:
              example:
                ok: true
                data:
                  messages:
                    - id: 2
                      conversation_id: 1
                      sender_type: bot
                      sender_id: 1
                      layers:
                        summary: Blog post ready.
                        thinking: Wrote 500 words.
                        status:
                          phase: complete
                          progress: 1.0
                      created_at: "2026-02-28T10:05:01Z"
                    - id: 1
                      conversation_id: 1
                      sender_type: user
                      sender_id: 1
                      layers:
                        summary: Write a blog post
                      created_at: "2026-02-28T10:05:00Z"
                  has_more: false
  /api/v1/conversations/{id}/participants:
    post:
      summary: Add a participant
      operationId: addParticipant
      security:
        - userAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entity_id:
                  type: integer
                  format: int64
                  deprecated: true
                  description: Legacy internal entity ID.
                entity_public_id:
                  type: string
                  format: uuid
                  description: Preferred participant public UUID.
                role:
                  type: string
                  enum: [member, admin, observer]
      responses:
        '201':
          description: Participant added
  /api/v1/friends:
    get:
      summary: List friends
      operationId: listFriends
      security:
        - userAuth: []
      parameters:
        - name: entity_id
          in: query
          schema:
            type: integer
            deprecated: true
          description: Legacy acting entity ID.
        - name: public_id
          in: query
          schema:
            type: string
            format: uuid
          description: Preferred acting entity public UUID.
      responses:
        '200':
          description: Friend entities
  /api/v1/friends/requests:
    get:
      summary: List friend requests
      operationId: listFriendRequests
      security:
        - userAuth: []
      parameters:
        - name: public_id
          in: query
          schema:
            type: string
            format: uuid
          description: Preferred acting entity public UUID.
        - name: entity_id
          in: query
          schema:
            type: integer
            deprecated: true
          description: Legacy acting entity ID.
      responses:
        '200':
          description: Friend requests
    post:
      summary: Create friend request
      operationId: createFriendRequest
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source_public_id:
                  type: string
                  format: uuid
                  description: Preferred acting entity public UUID.
                target_public_id:
                  type: string
                  format: uuid
                  description: Preferred friend target public UUID.
                source_entity_id:
                  type: integer
                  format: int64
                  deprecated: true
                target_entity_id:
                  type: integer
                  format: int64
                  deprecated: true
                message:
                  type: string
      responses:
        '201':
          description: Friend request created
  /api/v1/presence/batch:
    post:
      summary: Batch presence lookup
      operationId: batchPresence
      security:
        - userAuth: []
        - botAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                public_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Preferred entity public UUIDs.
                entity_ids:
                  type: array
                  deprecated: true
                  items:
                    type: integer
                    format: int64
                  description: Legacy internal entity IDs.
      responses:
        '200':
          description: Presence maps keyed by legacy IDs and public UUIDs
          content:
            application/json:
              example:
                ok: true
                data:
                  presence:
                    "1": true
                  presence_by_public_id:
                    419407cb-97a4-4d0f-a233-91ffc681d001: true
  /api/v1/messages/send:
    post:
      summary: Send a message
      operationId: sendMessage
      description: |
        Send a message to a conversation. Works for both users and bots.
        The server persists the message and broadcasts it to all participants via WebSocket.
      security:
        - userAuth: []
        - botAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [layers]
              properties:
                conversation_id:
                  type: integer
                  format: int64
                  description: Legacy internal conversation ID. Prefer conversation_public_id.
                conversation_public_id:
                  type: string
                  format: uuid
                  description: Preferred public conversation UUID.
                content_type:
                  type: string
                  enum: [text, markdown, code, image, audio, video, file, artifact, system, task_handover]
                layers:
                  $ref: '#/components/schemas/MessageLayers'
                attachments:
                  type: array
                  items:
                    $ref: '#/components/schemas/Attachment'
                stream_id:
                  type: string
                  description: Groups related messages in a stream
                mentions:
                  type: array
                  items:
                    type: integer
                    format: int64
                  description: Legacy internal entity IDs to mention.
                mention_public_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Preferred structured mention public UUIDs.
                mention_refs:
                  type: array
                  items:
                    $ref: '#/components/schemas/MentionRef'
                  description: Structured mention references selected by the client.
                assigned_public_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Subset of mentioned public IDs assigned to act. Empty array means mention-only.
                reply_to:
                  type: integer
                  format: int64
                  description: Message ID this message replies to.
            examples:
              simple:
                summary: Simple text message
                value:
                  conversation_id: 1
                  layers:
                    summary: Hello! How can I help?
              full:
                summary: Message with all layers
                value:
                  conversation_id: 1
                  layers:
                    summary: Blog post outline ready for review.
                    thinking: Analyzed the topic, identified 3 sections.
                    status:
                      phase: drafting
                      progress: 0.5
                      text: Outline complete
                    data:
                      outline:
                        - Why Agent-Native IM
                        - Architecture
                        - Use Cases
                    interaction:
                      type: approval
                      prompt: Approve this outline?
                      options:
                        - label: Approve
                          value: approve
                        - label: Revise
                          value: revise
      responses:
        '201':
          description: Message sent and persisted
  /api/v1/files/upload:
    post:
      summary: Upload a protected conversation file
      operationId: uploadFile
      security:
        - userAuth: []
        - botAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                conversation_id:
                  type: integer
                  format: int64
      responses:
        '201':
          description: File uploaded
  /api/v1/releases:
    get:
      summary: List release notes
      operationId: listReleases
      security:
        - userAuth: []
        - botAuth: []
      parameters:
        - name: channel
          in: query
          schema:
            type: string
            default: production
        - name: component
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Release notes with current user's unread count.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReleaseListResponse'
  /api/v1/releases/latest:
    get:
      summary: Get latest release note
      operationId: getLatestRelease
      security:
        - userAuth: []
        - botAuth: []
      parameters:
        - name: channel
          in: query
          schema:
            type: string
            default: production
      responses:
        '200':
          description: Latest release note for the selected channel.
          content:
            application/json:
              example:
                ok: true
                data:
                  release:
                    id: 1
                    version: "2026.5.14"
                  unread_count: 1
  /api/v1/releases/{id}/read:
    post:
      summary: Mark a release note as read
      operationId: markReleaseRead
      security:
        - userAuth: []
        - botAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Release read marker upserted.
  /api/v1/updates:
    get:
      summary: Long polling for new messages
      operationId: getUpdates
      description: |
        Returns new messages for the authenticated bot. If no messages are available,
        holds the connection open until a message arrives or timeout.
      security:
        - botAuth: []
      parameters:
        - name: offset
          in: query
          description: Only return messages with id > offset
          schema:
            type: integer
            default: 0
        - name: timeout
          in: query
          description: Max seconds to wait (1-60)
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Array of new messages (may be empty on timeout)
          content:
            application/json:
              example:
                ok: true
                data:
                  - id: 5
                    conversation_id: 1
                    sender_type: user
                    sender_id: 1
                    layers:
                      summary: Please write a report
                    created_at: "2026-02-28T10:10:00Z"
  /api/v1/ws:
    get:
      summary: WebSocket connection
      operationId: websocket
      description: |
        Upgrade to WebSocket for real-time bidirectional messaging.
        Pass token as query parameter.

        **Inbound message types** (client → server):
        - `message.send` — send a message (same format as REST send)
        - `ping` — keepalive

        **Outbound message types** (server → client):
        - `message.new` — new message in a conversation you participate in
        - `stream.start` — bot started streaming (ephemeral)
        - `stream.delta` — streaming progress update (ephemeral)
        - `stream.end` — streaming complete (persisted)
        - `pong` — keepalive response
        - `error` — error notification

        **Streaming via WebSocket:**
        ```json
        // Start stream
        {"type": "message.send", "data": {
          "conversation_id": 1, "stream_id": "task-1",
          "stream_type": "start",
          "layers": {"status": {"phase": "starting", "progress": 0}}
        }}
        // Delta (not persisted)
        {"type": "message.send", "data": {
          "conversation_id": 1, "stream_id": "task-1",
          "stream_type": "delta",
          "layers": {"status": {"phase": "working", "progress": 0.5}}
        }}
        // End (persisted)
        {"type": "message.send", "data": {
          "conversation_id": 1, "stream_id": "task-1",
          "stream_type": "end",
          "layers": {"summary": "Task complete."}
        }}
        ```
      parameters:
        - name: token
          in: query
          required: true
          description: JWT token (user) or API token (bot)
          schema:
            type: string