# internal/handler/router.go
package handler

import (
	"log/slog"
	"net/http"
	"strings"
	"sync"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/wzfukui/agent-native-im/internal/auth"
	"github.com/wzfukui/agent-native-im/internal/config"
	"github.com/wzfukui/agent-native-im/internal/filestore"
	"github.com/wzfukui/agent-native-im/internal/middleware"
	"github.com/wzfukui/agent-native-im/internal/model"
	"github.com/wzfukui/agent-native-im/internal/push"
	"github.com/wzfukui/agent-native-im/internal/store"
	"github.com/wzfukui/agent-native-im/internal/webhook"
	"github.com/wzfukui/agent-native-im/internal/ws"
)

type AuthHelper struct {
	Secret   string
	TokenTTL time.Duration
}

func (a *AuthHelper) GenerateToken(entityID int64, entityType model.EntityType) (string, error) {
	return auth.GenerateTokenWithTTL(a.Secret, entityID, entityType, a.TokenTTL)
}

type Server struct {
	Config        *config.Config
	Store         store.Store
	Hub           *ws.Hub
	Webhook       *webhook.Deliverer
	Auth          *AuthHelper
	FileStore     filestore.FileStore
	Push          *push.Sender
	RotationLocks sync.Map
}

func NewRouter(s *Server) *gin.Engine {
	r := gin.Default()
	r.Use(middleware.RequestID())
	r.Use(middleware.JSONLogger())
	r.Use(securityHeaders())
	r.Use(corsMiddleware())

	// Create rate limiters
	rateLimiters := middleware.CreateRateLimiters()

	v1 := r.Group("/api/v1")
	{
		// Public
		v1.GET("/ping", HandlePing)
		v1.GET("/health", s.HandleHealthCheck)
		v1.GET("/skill-template", HandleSkillTemplate)
		v1.GET("/onboarding-guide", s.HandleOnboardingGuide)

		// Auth endpoints with strict rate limiting
		v1.POST("/auth/login", rateLimiters["login"].Middleware(), s.HandleLogin)
		v1.POST("/auth/register", rateLimiters["register"].Middleware(), s.HandleRegister)
		v1.GET("/auth/1pass/config", s.HandleOnePassConfig)
		v1.POST("/auth/1pass/login", rateLimiters["login"].Middleware(), s.HandleOnePassLogin)

		// Public push key endpoint (no auth needed)
		v1.GET("/push/vapid-key", s.HandleGetVAPIDKey)
		v1.GET("/public/bots/:identifier", s.HandleGetPublicBot)
		v1.POST("/public/bots/:identifier/session", rateLimiters["login"].Middleware(), s.HandleCreatePublicBotSession)

		// Authenticated (JWT or permanent API key)
		authed := v1.Group("")
		authed.Use(auth.EntityAuth(s.Config.JWTSecret, s.Store))
		authed.Use(middleware.Audit(s.Store))
		{
			authed.GET("/me", s.HandleMe)
			authed.POST("/auth/refresh", s.HandleRefreshToken)
			authed.POST("/auth/logout", s.HandleLogout)

			full := authed.Group("")
			full.Use(auth.RequireFullAuth())
			{
				// User management
				full.GET("/me/auth-methods", s.HandleAuthMethods)
				full.PUT("/me", s.HandleUpdateProfile)
				full.PUT("/me/password", s.HandleChangePassword)
				full.POST("/me/external-identities/1pass/link", rateLimiters["login"].Middleware(), s.HandleLinkOnePass)
				full.DELETE("/me/external-identities/:id", s.HandleUnlinkExternalIdentity)

				full.GET("/me/devices", s.HandleListDevices)
				full.DELETE("/me/devices/:deviceId", s.HandleKickDevice)

				// Admin-only endpoints
				admin := full.Group("")
				admin.Use(auth.RequireAdmin(s.Store, s.Config.AdminUser))
				{
					admin.POST("/admin/users", s.HandleCreateUser)
					admin.GET("/admin/users", s.HandleAdminListUsers)
					admin.PUT("/admin/users/:id", s.HandleAdminUpdateUser)
					admin.DELETE("/admin/users/:id", s.HandleAdminDeleteUser)
					admin.GET("/admin/stats", s.HandleAdminStats)
					admin.GET("/admin/conversations", s.HandleAdminListConversations)
					admin.GET("/admin/audit-logs", s.HandleAdminListAuditLogs)
					admin.POST("/admin/reset-password", s.HandleAdminResetPassword)
					admin.PATCH("/admin/feedback/:id", s.HandleAdminUpdateFeedback)
				}
				// Entity management (user-only at handler level)
				full.POST("/entities", s.HandleCreateEntity)
				full.GET("/entities", s.HandleListEntities)
				full.GET("/entities/discover", rateLimiters["discover"].Middleware(), s.HandleSearchDiscoverableEntities)
				full.GET("/entities/search", s.HandleSearchEntities)
				full.PUT("/entities/:id", s.HandleUpdateEntity)
				full.DELETE("/entities/:id", s.HandleDeleteEntity)
				full.POST("/entities/:id/approve", s.HandleApproveConnection)
				full.GET("/entities/:id/status", s.HandleEntityStatus)
				full.GET("/entities/:id/credentials", s.HandleGetCredentials)
				full.GET("/entities/:id/self-check", s.HandleEntitySelfCheck)
				full.GET("/entities/:id/diagnostics", s.HandleEntityDiagnostics)
				full.POST("/entities/:id/regenerate-token", s.HandleRegenerateEntityToken)
				full.POST("/entities/:id/reactivate", s.HandleReactivateEntity)
				full.POST("/presence/batch", s.HandleBatchPresence)
				full.POST("/bots/:id/access-links", s.HandleCreateBotAccessLink)
				full.GET("/bots/:id/access-links", s.HandleListBotAccessLinks)
				full.DELETE("/bot-access-links/:id", s.HandleDeleteBotAccessLink)

				// Friends
				full.GET("/friends", s.HandleListFriends)
				full.GET("/friends/requests", s.HandleListFriendRequests)
				full.GET("/inbox/snapshot", s.HandleInboxSnapshot)
				full.POST("/friends/requests", s.HandleCreateFriendRequest)
				full.POST("/friends/requests/:id/accept", s.HandleAcceptFriendRequest)
				full.POST("/friends/requests/:id/reject", s.HandleRejectFriendRequest)
				full.POST("/friends/requests/:id/cancel", s.HandleCancelFriendRequest)
				full.DELETE("/friends/:entityId", s.HandleDeleteFriend)
				full.GET("/notifications", s.HandleListNotifications)
				full.POST("/notifications/:id/read", s.HandleMarkNotificationRead)
				full.POST("/notifications/read-all", s.HandleMarkAllNotificationsRead)
				full.POST("/feedback", s.HandleCreateFeedback)
				full.GET("/feedback", s.HandleListFeedback)
				full.GET("/feedback/:id", s.HandleGetFeedback)
				full.POST("/feedback/:id/comments", s.HandleCreateFeedbackComment)
				full.GET("/releases", s.HandleListReleases)
				full.GET("/releases/latest", s.HandleLatestRelease)
				full.POST("/releases/:id/read", s.HandleMarkReleaseRead)

				// Webhook management
				full.POST("/webhooks", s.HandleCreateWebhook)
				full.GET("/webhooks", s.HandleListWebhooks)
				full.DELETE("/webhooks/:id", s.HandleDeleteWebhook)

				// Conversations
				full.POST("/conversations", s.HandleCreateConversation)
				full.GET("/conversations", s.HandleListConversations)
				full.GET("/conversations/:id", s.HandleGetConversation)
				full.GET("/conversations/by-public-id/:publicId", s.HandleGetConversationByPublicID)
				full.GET("/conversations/public/:publicId", s.HandleGetConversationByPublicIDCompat)
				full.PUT("/conversations/:id", s.HandleUpdateConversation)

				// Participants & lifecycle
				full.POST("/conversations/:id/participants", s.HandleAddParticipant)
				full.DELETE("/conversations/:id/participants/:entityId", s.HandleRemoveParticipant)
				full.PUT("/conversations/:id/subscription", s.HandleUpdateSubscription)
				full.POST("/conversations/:id/read", s.HandleMarkAsRead)
				full.POST("/conversations/:id/leave", s.HandleLeaveConversation)
				full.POST("/conversations/:id/archive", s.HandleArchiveConversation)
				full.POST("/conversations/:id/unarchive", s.HandleUnarchiveConversation)
				full.POST("/conversations/:id/pin", s.HandlePinConversation)
				full.POST("/conversations/:id/unpin", s.HandleUnpinConversation)

				// Progress (transient, no DB write)
				full.POST("/conversations/:id/progress", s.HandleSendProgress)

				// Typing indicator (transient, no DB write)
				full.POST("/conversations/:id/typing", s.HandleSendTyping)

				// Messages (with rate limiting)
				full.POST("/messages/send", rateLimiters["message"].Middleware(), s.HandleSendMessage)
				full.GET("/messages/:id", s.HandleGetMessage)
				full.DELETE("/messages/:id", s.HandleRevokeMessage)
				full.PUT("/messages/:id", s.HandleEditMessage)
				full.POST("/messages/:id/respond", s.HandleInteractionResponse)
				full.POST("/messages/:id/reactions", s.HandleToggleReaction)
				full.GET("/conversations/:id/messages", s.HandleListMessages)
				full.GET("/conversations/:id/search", s.HandleSearchMessages)
				full.GET("/messages/search", s.HandleGlobalSearchMessages)

				// Invite links
				full.POST("/conversations/:id/invite", s.HandleCreateInviteLink)
				full.GET("/conversations/:id/invites", s.HandleListInviteLinks)
				full.DELETE("/invites/:id", s.HandleDeleteInviteLink)

				// File upload (with rate limiting)
				full.POST("/files/upload", rateLimiters["file"].Middleware(), s.HandleFileUpload)

				// Push notifications
				full.POST("/push/subscribe", s.HandleRegisterPush)
				full.POST("/push/unsubscribe", s.HandleUnregisterPush)

				// Invite join
				full.GET("/invite/:code", s.HandleGetInviteInfo)
				full.POST("/invite/:code/join", s.HandleJoinViaInvite)

				// Tasks
				full.POST("/conversations/:id/tasks", s.HandleCreateTask)
				full.GET("/conversations/:id/tasks", s.HandleListTasks)
				full.GET("/tasks/:taskId", s.HandleGetTask)
				full.PUT("/tasks/:taskId", s.HandleUpdateTask)
				full.DELETE("/tasks/:taskId", s.HandleDeleteTask)

				// Memories
				full.GET("/conversations/:id/memories", s.HandleListMemories)
				full.POST("/conversations/:id/memories", s.HandleUpsertMemory)
				full.DELETE("/conversations/:id/memories/:memId", s.HandleDeleteMemory)

				// Change Requests
				full.POST("/conversations/:id/change-requests", s.HandleCreateChangeRequest)
				full.GET("/conversations/:id/change-requests", s.HandleListChangeRequests)
				full.POST("/conversations/:id/change-requests/:reqId/approve", s.HandleApproveChangeRequest)
				full.POST("/conversations/:id/change-requests/:reqId/reject", s.HandleRejectChangeRequest)

				// Long polling
				full.GET("/updates", s.HandleUpdates)
			}
		}

		// WebSocket (auth via Authorization header, subprotocol, or cookie)
		v1.GET("/ws", s.HandleWS)

		// Authenticated file serving (replaces public static)
		if s.FileStore != nil {
			r.GET("/avatars/*filename", s.HandleAvatarDownload)
			r.GET("/avatar-files/*filename", s.HandleAvatarDownload)
			fileGroup := r.Group("/files")
			fileGroup.Use(auth.EntityAuth(s.Config.JWTSecret, s.Store))
			fileGroup.GET("/*filename", s.HandleFileDownload)
		}
	}

	return r
}

func corsMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		origin := c.GetHeader("Origin")

		c.Header("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE, OPTIONS")
		c.Header("Access-Control-Allow-Headers", "Origin, Content-Type, Authorization")
		c.Header("Access-Control-Max-Age", "86400")

		if origin != "" && isAllowedBrowserOrigin(origin) {
			// Allowed origin: enable credentials
			c.Header("Access-Control-Allow-Origin", origin)
			c.Header("Access-Control-Allow-Credentials", "true")
		} else if origin == "" {
			// No origin header (same-origin or non-browser): allow
			c.Header("Access-Control-Allow-Origin", "*")
		} else {
			// Rejected origin: no CORS headers
			slog.Warn("CORS rejected", "origin", origin)
		}

		if strings.ToUpper(c.Request.Method) == "OPTIONS" {
			c.AbortWithStatus(http.StatusNoContent)
			return
		}

		c.Next()
	}
}

// securityHeaders adds security-related HTTP headers to all responses
func securityHeaders() gin.HandlerFunc {
	return func(c *gin.Context) {
		// Prevent clickjacking attacks
		c.Header("X-Frame-Options", "DENY")

		// Prevent MIME type sniffing
		c.Header("X-Content-Type-Options", "nosniff")

		// Enable XSS protection in older browsers
		c.Header("X-XSS-Protection", "1; mode=block")

		// Enforce HTTPS (only for production)
		if !strings.Contains(c.Request.Host, "localhost") && !strings.Contains(c.Request.Host, "127.0.0.1") {
			c.Header("Strict-Transport-Security", "max-age=31536000; includeSubDomains")
		}

		// Content Security Policy - adjust based on your needs
		// Allow self, data URIs for images (SVG avatars), and specific WebSocket origins
		csp := "default-src 'self'; " +
			"img-src 'self' data: https:; " +
			"script-src 'self' 'unsafe-inline' 'unsafe-eval'; " + // unsafe-eval needed for React dev tools
			"style-src 'self' 'unsafe-inline'; " +
			"connect-src 'self' https: wss: http://localhost:* http://127.0.0.1:* http://192.168.44.43:* ws://localhost:* ws://127.0.0.1:* ws://192.168.44.43:*; " +
			"font-src 'self' data:; " +
			"object-src 'none'; " +
			"frame-ancestors 'none'"
		c.Header("Content-Security-Policy", csp)

		// Referrer policy
		c.Header("Referrer-Policy", "strict-origin-when-cross-origin")

		// Permissions policy (replaces Feature-Policy)
		c.Header("Permissions-Policy", "camera=(), microphone=(), geolocation=()")

		c.Next()
	}
}